fix bug IN delete VospPlan

+++ Form1.cs
@@ -1,5 +1,6 @@
 using ArchiveFund;
 using MySql.Data.MySqlClient;
+using Mysqlx.Crud;
 using System.Text;
 
 namespace Praktika01Uvarov
@@ -294,28 +295,32 @@ JOIN Educational_work_plan ON Educational_work_plan.Number_plan = `Event`.fk_Num
                         int idNumberPlan = Convert.ToInt32(dataGridView1.Rows[indRow].Cells[0].Value);
                         if (dataGridView2.RowCount != 0 || dataGridView4.RowCount != 0)
                         {
+                            if(dataGridView2.RowCount != 0 && indRow >= dataGridView2.RowCount)
+                            {
                             int idNumberPlanEvent = Convert.ToInt32(dataGridView2.Rows[indRow].Cells["EventIDNumber"].Value);
                             sqlCommand = $"DELETE FROM `event` WHERE fk_Number_plan = {idNumberPlanEvent.ToString()};";
                             cmd = new(sqlCommand, conn);
                             cmd.ExecuteNonQuery();
                             fillTable2();
-                            sqlCommand = $"DELETE FROM `inviting_participants` WHERE fk_Number_plan = {idNumberPlan.ToString()};";
-                            cmd = new(sqlCommand, conn);
-                            cmd.ExecuteNonQuery();
+                            }
+                            if (dataGridView4.RowCount != 0 && indRow >= dataGridView4.RowCount)
+                            {
+                                sqlCommand = $"DELETE FROM `inviting_participants` WHERE fk_Number_plan = {idNumberPlan.ToString()};";
+                                cmd = new(sqlCommand, conn);
+                                cmd.ExecuteNonQuery();
+                                int idPlayer = Convert.ToInt32(dataGridView4.Rows[indRow].Cells[1].Value);
+                                sqlCommand = $"DELETE FROM Invited_participants WHERE Code_player = '{idPlayer.ToString()}';";
+                                cmd = new(sqlCommand, conn);
+                                cmd.ExecuteNonQuery();
+                                fillTable4();
+                            }
                             sqlCommand = $"DELETE FROM Educational_work_plan WHERE Number_plan = {idNumberPlan.ToString()}";
                             cmd = new(sqlCommand, conn);
                             cmd.ExecuteNonQuery();
-                            fillTable();
-                        }
-                        else
-                        {
-                            sqlCommand = $"DELETE FROM Educational_work_plan WHERE Number_plan = {idNumberPlan.ToString()}";
-                            cmd = new(sqlCommand, conn);
-                            cmd.ExecuteNonQuery();
-                            fillTable();
                         }
                     }
                 }
+                fillTable();
             }
             catch (Exception ex)
             {
